+++ backends/firebase/kmpauth-firebase-facebook/src/jsMain/kotlin/com/mmk/kmpauth/firebase/facebook/FirebaseFacebookSignInState.js.kt
@@ -7,6 +7,7 @@ import com.mmk.kmpauth.core.KMPAuthInternalApi
 import com.mmk.kmpauth.core.LaunchingSignInState
 import com.mmk.kmpauth.core.SignInState
 import com.mmk.kmpauth.core.logger.currentLogger
+import com.mmk.kmpauth.facebook.FacebookLoginTracking
 import com.mmk.kmpauth.facebook.FacebookSignInRequestScope
 import dev.gitlive.firebase.auth.FirebaseUser
 
@@ -15,6 +16,7 @@ import dev.gitlive.firebase.auth.FirebaseUser
 public actual fun rememberFirebaseFacebookSignInState(
     requestScopes: List<FacebookSignInRequestScope>,
     linkAccount: Boolean,
+    loginTracking: FacebookLoginTracking,
     onResult: (Result<FirebaseUser?>) -> Unit,
 ): SignInState {
     val scope = rememberCoroutineScope()